+++ spec/fixtures/filters/post/post_filter.rb
@@ -1,0 +1,7 @@
+module Schleuder
+  module Filters
+    def self.post_filter(list, mail)
+      nil
+    end
+  end
+end