+++ spec/fixtures/filters/pre/pre_filter.rb
@@ -1,0 +1,7 @@
+module Schleuder
+  module Filters
+    def self.pre_filter(list, mail)
+      nil
+    end
+  end
+end