+++ CHANGELOG.md
@@ -10,6 +10,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * The output of the keywords 'X-ADD-KEY' and 'X-DELETE-KEY' now also show the "oneline"-format to represent keys (which includes fingerprint, primary email-address, date of generation and possible expiry). (#295)
 * In the response to 'X-ADD-KEY', differentiate between 'newly imported' and 'updated' keys.
 
+### Added
+
+* Enable to load external filters, similar to how we allow external plugins. (#200)
 
 ## [3.2.3] / 2018-05-14
 
@@ -34,7 +37,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Check environment variable if code coverage check should be executed. (#342)
 * Transform GPG fingerprints to upper case before saving to database. (#327)
 * CLI-commands that (potentially) change data now remind the system admin to check file system permission if the command was run with root privileges. (#326)
-
 
 ## [3.2.2] / 2018-02-06